+++ src/components/msg/vesting/create_periodic_vesting_account/index.test.tsx
@@ -1,0 +1,49 @@
+import React from 'react';
+import { RecoilRoot } from 'recoil';
+import renderer from 'react-test-renderer';
+import { MockTheme } from '@tests/utils';
+import { MsgCreatePeriodicVestingAccount } from '@models';
+import CreatePeriodicVestingAccount from '.';
+
+// ==================================
+// mocks
+// ==================================
+
+jest.mock('@components', () => ({
+  Name: (props) => <div id="Name" {...props} />,
+}));
+
+jest.mock('next-translate/Trans', () => (
+  (props) => <div id="Trans" {...props} />
+));
+
+// ==================================
+// unit tests
+// ==================================
+describe('screen: TransactionDetails/Grant', () => {
+  it('matches snapshot', () => {
+    const message = new MsgCreatePeriodicVestingAccount({
+      category: 'vesting',
+      type: 'MsgGrant',
+      toAddress: 'toAddress',
+      fromAddress: 'fromAddress',
+    });
+    const component = renderer.create(
+      <RecoilRoot>
+        <MockTheme>
+          <CreatePeriodicVestingAccount
+            message={message}
+          />
+        </MockTheme>
+      </RecoilRoot>,
+    );
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+
+    expect(component.root.findByProps({ id: 'Trans' }).props.i18nKey).toEqual('message_contents:MsgCreatePeriodicVestingAccount');
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+});