+++ src/screens/proposal_details/components/deposits/components/desktop/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import classnames from 'classnames';
 import useTranslation from 'next-translate/useTranslation';
-import numeral from 'numeral';
 import {
   Table,
   TableHead,
@@ -10,6 +9,7 @@ import {
   TableBody,
 } from '@material-ui/core';
 import { AvatarName } from '@components';
+import { formatNumber } from '@utils/format_token';
 import { columns } from './utils';
 import { ItemType } from '../../types';
 
@@ -36,7 +36,7 @@ const Desktop: React.FC<{
           )}
         </>
       ),
-      amount: `${numeral(x.amount.value).format(x.amount.format)} ${x.amount.denom.toUpperCase()}`,
+      amount: `${formatNumber(x.amount.value, x.amount.exponent)} ${x.amount.displayDenom.toUpperCase()}`,
     });
   });