+++ src/screens/transaction_details/hooks.ts
@@ -7,7 +7,7 @@ import {
   useTransactionDetailsQuery,
   TransactionDetailsQuery,
 } from '@graphql/types';
-import { formatDenom } from '@utils/format_denom';
+import { formatToken } from '@utils/format_token';
 import { convertMsgsToModels } from '@msg';
 import {
   TransactionState,
@@ -23,9 +23,10 @@ export const useTransactionDetails = () => {
       height: 0,
       timestamp: '',
       fee: {
-        value: 0,
-        denom: '',
-        format: '',
+        value: '0',
+        displayDenom: '',
+        baseDenom: '',
+        exponent: 0,
       },
       gasUsed: 0,
       gasWanted: 0,
@@ -90,7 +91,7 @@ export const useTransactionDetails = () => {
         hash: data.transaction[0].hash,
         height: data.transaction[0].height,
         timestamp: data.transaction[0].block.timestamp,
-        fee: formatDenom(feeAmount.amount, feeAmount.denom),
+        fee: formatToken(feeAmount.amount, feeAmount.denom),
         gasUsed: data.transaction[0].gasUsed,
         gasWanted: data.transaction[0].gasWanted,
         success,