+++ src/screens/validator_details/components/voting_power/index.tsx
@@ -14,17 +14,21 @@ import { VotingPowerType } from '../../types';
 const VotingPower: React.FC<{
   className?: string;
   data: VotingPowerType;
+  status: number;
 }> = ({
   className,
   data,
+  status,
 }) => {
   const { t } = useTranslation('validators');
-  const votingPowerPercent = numeral((
-    data.self / data.overall.value) * 100);
+  const votingPowerPercent = status === 3 ? numeral((
+    data.self / numeral(data.overall.value).value()) * 100) : numeral(0);
 
   const classes = useStyles(votingPowerPercent.format(0, Math.floor));
 
-  const votingPower = numeral(data.self).format('0,0');
+  const votingPower = status === 3 ? numeral(data.self).format('0,0') : '0';
+  const selfDelegatedAmount = status === 3 ? numeral(data.selfDelegate.value).format('0,0') : '0';
+  const selfDelegatePercent = status === 3 ? `${numeral(data.selfDelegatePercent).format('0.[00]')}%` : '0%';
 
   return (
     <Box className={classnames(className, classes.root)}>
@@ -69,10 +73,10 @@ const VotingPower: React.FC<{
           {t('selfDelegatedTokens')}
         </Typography>
         <Typography variant="body1" className="value">
-          {numeral(data.selfDelegate.value).format('0,0')}
+          {selfDelegatedAmount}
           {' '}
           (
-          {`${numeral(data.selfDelegatePercent).format('0.[00]')}%`}
+          {selfDelegatePercent}
           )
         </Typography>
       </div>